clawpatch_output_vtk: avoid possible warnings

+++ src/patches/clawpatch/fclaw2d_clawpatch_output_vtk.c
@@ -334,7 +334,7 @@ write_offsets_cb (fclaw2d_domain_t * domain, fclaw2d_patch_t * patch,
     if (s->fits32)
     {
         int32_t *idata = (int32_t *) s->buf;
-        int32_t k = connectivity_size * (cbefore + 1);
+        int32_t k = connectivity_size * (int32_t) (cbefore + 1);
         for (c = 0; c < s->cells_per_patch; k += connectivity_size, ++c)
         {
             *idata++ = k;
@@ -417,7 +417,7 @@ write_patchno_cb (fclaw2d_domain_t * domain, fclaw2d_patch_t * patch,
 
     if (s->fits32)
     {
-        const int32_t igpno = (int32_t) gpno;
+        const int32_t igpno = (const int32_t) gpno;
         int32_t *idata = (int32_t *) s->buf;
         for (c = 0; c < s->cells_per_patch; ++c)
         {
